Add tests for super-user and self deletion, refactor user-dashboard.js (#305)
The tests are fairly straightforward. The refactor consisted of moving
the user and group creation to a 'before' step, rather than deleting and
recreating the user at the end of each test. This shortens the test run
by a tiny bit and also makes it easier to understand what's going on
where.